# gatewayservice/openapi.yaml
openapi: 3.0.0
info:
  title: Gateway API
  description: Gateway OpenApi specification.
  version: 0.1.9
servers:
  - url: http://localhost:8000
    description: Development server
  - url: http://SOMEIP:8000
    description: Production server
paths:
  /adduser:
    post:
      summary: Adds a new user to the database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: The user ID.
                password:
                  type: string
                  description: The user password.
      responses:
        '200':    # status code
          description: OK
        '400':    # status code
          description: The user cannot be added
          content:
            application/json:
              schema: 
                type: object
                properties:
                  error:
                    type: string
                    description: Information about the error.
  /health:
    get:
      summary: the Healthcheck of the service.
      responses:
        '200':    # status code
          description: The status Code
          content:
            application/json:
              schema: 
                type: object
                properties:
                  status:
                    type: string
                    description: OK
  /login:
    post:
      summary: Logs the user into the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: The user ID.
                password:
                  type: string
                  description: The user password.
      responses:
        '200':    # status code
          description: A JSON array of user names
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: The user token.
                  username:
                    type: string
                    description: The username.
                  createdAt:
                    type: string ## We can use patterns to validate it. See more: https://swagger.io/docs/specification/data-models/data-types/
                    description: The username.
        '401':    # status code
          description: The user cannot be added
          content:
            application/json:
              schema: 
                type: object
                properties:
                  error:
                    type: string
                    description: Invalid Credentials.
        '500':    # status code
          description: The user cannot be added
          content:
            application/json:
              schema: 
                type: object
                properties:
                  error:
                    type: string
                    description: Information about the error.